refactor(chat): unify convertToThreadMessageLike into shared module
Extract the duplicated message conversion logic into a platform-agnostic
messageTransform.ts. The native file re-exports from it and keeps only
transformMessageLike (which needs react-native internals). The history
adapter is now generic to avoid @assistant-ui/react type imports.
